added check for path to save_to_disk func. Covered it with tests

+++ lesson_02/ht_template/job1/dal/local_disk.py
@@ -7,8 +7,10 @@ import shutil
 # if sales directory exists, it will be deleted
 # then creates new directory sales/date and writes file there
 def save_to_disk(json_content: List[Dict[str, Any]], path: str, date: str) -> None:
+    if os.path.join('raw', 'sales') not in path:
+        raise ValueError
+
     sales_path = os.path.abspath(os.path.join(path, os.pardir))
-    print(sales_path)
     if os.path.exists(sales_path) and "sales" in str(sales_path):
         shutil.rmtree(sales_path)
